+++ crates/sui-indexer-alt-framework/src/pipeline/concurrent/commit_watermark.rs
@@ -158,22 +158,22 @@ pub(super) fn commit_watermark<H: Handler + 'static>(
                     metrics
                         .watermark_epoch
                         .with_label_values(&[H::NAME])
-                        .set(watermark.epoch_hi_inclusive);
+                        .set(watermark.epoch_hi_inclusive as i64);
 
                     metrics
                         .watermark_checkpoint
                         .with_label_values(&[H::NAME])
-                        .set(watermark.checkpoint_hi_inclusive);
+                        .set(watermark.checkpoint_hi_inclusive as i64);
 
                     metrics
                         .watermark_transaction
                         .with_label_values(&[H::NAME])
-                        .set(watermark.tx_hi);
+                        .set(watermark.tx_hi as i64);
 
                     metrics
                         .watermark_timestamp_ms
                         .with_label_values(&[H::NAME])
-                        .set(watermark.timestamp_ms_hi_inclusive);
+                        .set(watermark.timestamp_ms_hi_inclusive as i64);
 
                     debug!(
                         pipeline = H::NAME,
@@ -221,17 +221,17 @@ pub(super) fn commit_watermark<H: Handler + 'static>(
                                 metrics
                                     .watermark_epoch_in_db
                                     .with_label_values(&[H::NAME])
-                                    .set(watermark.epoch_hi_inclusive);
+                                    .set(watermark.epoch_hi_inclusive as i64);
 
                                 metrics
                                     .watermark_transaction_in_db
                                     .with_label_values(&[H::NAME])
-                                    .set(watermark.tx_hi);
+                                    .set(watermark.tx_hi as i64);
 
                                 metrics
                                     .watermark_timestamp_in_db_ms
                                     .with_label_values(&[H::NAME])
-                                    .set(watermark.timestamp_ms_hi_inclusive);
+                                    .set(watermark.timestamp_ms_hi_inclusive as i64);
                             }
                             Ok(false) => {}
                         }
